# reference/marketplace-openapi.yaml
openapi: 3.0.2
...
info:
  description: |
    # Overview

    Noona provides a powerful REST(y) API that focuses on simplicity and clarity.

    - All API requests must be made over HTTPS. Calls made over plain HTTP will be redirected to HTTPS.

    **Base Path**
    ```
    https://api.noona.is
    ```

    The Marketplace API allows various 'customer' interactions with the companies in the Noona ecosystem.

    Example user: **An individual seeking a haircut**.

  title: Noona Marketplace API
  version: dev
  x-logo:
    url: https://api.noona.is/docs/img/logo_black.svg
servers:
  - url: https://api.noona.is
...
paths:
...
  /v1/marketplace/companies:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
    get:
      summary: List companies
      description: Retrieves all companies accessible on the Noona marketplace.
      parameters:
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyFilter"
        - name: sort
          in: query
          description: "[Sorting](https://api.noona.is/docs/working-with-the-apis/sorting)"
          required: false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySort"
        - name: pagination
          in: query
          description: "[Pagination](https://api.noona.is/docs/working-with-the-apis/pagination)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pagination"
        - $ref: "#/components/parameters/search"
        - name: sort_by
          in: query
          description: Use sort instead.
          deprecated: true
          required: false
          schema:
            type: string
            enum: [popular, distance]
        - name: lat
          in: query
          required: false
          deprecated: true
          description: Use filter instead.
          schema:
            type: number
            format: double
        - name: lng
          in: query
          required: false
          deprecated: true
          description: Use filter instead.
          schema:
            type: number
            format: double
        - name: radius
          in: query
          required: false
          deprecated: true
          description: Use filter instead.
          schema:
            type: integer
            format: int64
            default: 100000
        - name: company_type_id
          in: query
          description: Use filter instead.
          required: false
          deprecated: true
          schema:
            type: string
        - name: enterprise_id
          in: query
          description: Use filter instead.
          deprecated: true
          required: false
          schema:
            type: string
            example: 7awdXawZoolkjad37a
        - name: limit
          in: query
          description: Use pagination object instead.
          deprecated: true
          required: false
          schema:
            type: integer
            format: int64
        - name: skip
          in: query
          description: Use pagination object instead.
          deprecated: true
          required: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Companies"
          description: Companies
      operationId: ListCompanies
      tags:
        - Marketplace Companies
...
  /v1/marketplace/companies/{company_id}/employees:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
      - name: sort
        in: query
        description: "[Sorting](https://api.noona.is/docs/working-with-the-apis/sorting)"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sort"
      - name: pagination
        in: query
        description: "[Pagination](https://api.noona.is/docs/working-with-the-apis/pagination)"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pagination"
    get:
      summary: List employees
      description: |
        Lists employees at a given company.
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: string
            example: "aw7da9wd8ua28a821"
            description: "Company ID"
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeFilter"
        - name: event_type_ids
          in: query
          deprecated: true
          description: Use filter object instead.
          schema:
            type: array
            example: ["te7da9wd8ua28a1ab", "xa7da9wd8ua01a134"]
            items:
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employees"
          description: Employees
      operationId: ListEmployees
      tags:
        - Employees
...
  /v1/marketplace/events:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
    get:
      security:
        - Marketplace-Authentication: []
      summary: List events
      description: |
        Lists events for marketplace user.
      parameters:
        - name: sort
          in: query
          description: "[Sorting](https://api.noona.is/docs/working-with-the-apis/sorting)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sort"
        - name: pagination
          in: query
          description: "[Pagination](https://api.noona.is/docs/working-with-the-apis/pagination)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pagination"
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventFilter"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Events"
          description: Events
      operationId: ListEvents
      tags:
        - Marketplace Events
    post:
      security:
        - Marketplace-Authentication: []
      summary: Create an event
      description: |
        Creates an event for marketplace user from a time slot reservation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Event"
                - $ref: "#/components/schemas/EventCreateOverrides"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
          description: Created Event
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateEventErrorOneOf"
      operationId: CreateEvent
      tags:
        - Marketplace Events
  /v1/marketplace/events/{event_id}:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
    get:
      security:
        - Marketplace-Authentication: []
      summary: Retrieve an event
      description: |
        Retrieves an event for marketplace user.
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            example: "dwawd8awudawd"
            description: Event ID
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
          description: Events
      operationId: GetEvent
      tags:
        - Marketplace Events
    post:
      security:
        - Marketplace-Authentication: []
      summary: Update an event
      description: |
        Updates an event for marketplace user. 

        If employee and space are not provided, new employee and space will be selected at random for the new time.

        If the intent is to only move between resources (employee or space) and not change the time, starts_at will have to be passed in as the current value.

        If the intent is to only change number of guests, starts_at will have to be passed in as the current value.
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            example: "dwawd8awudawd"
            description: Event ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
          description: Updated Event
      operationId: UpdateEvent
      tags:
        - Marketplace Events
...
  /v1/marketplace/user/verify_phone_number:
    post:
      summary: Verify phone number
      description: |
        Verifies a phone number by sending a unique code to it.

        The code is then used in conjuction with the phone number to create a new user or verify an existing one.

        This endpoint can also be used to verify phone numbers of users that have authenticated
        with external identity providers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPhoneNumberRequest"
      responses:
        "200":
          description: Verification number was sent to phone number.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPhoneNumberResult"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPhoneNumberError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPhoneNumberError"
      operationId: VerifyPhoneNumber
      tags:
        - Marketplace User
  /v1/marketplace/user/verified:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
    post:
      summary: Creates a new user
      description: |
        Creates a new user with a verified phone number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifiedPhone"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Created marketplace user
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: CreateVerifiedUser
      tags:
        - Marketplace User
  /v1/marketplace/user:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
    get:
      security:
        - Marketplace-Authentication: []
      summary: Get user
      description: Gets current marketplace user, inferred from token.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Marketplace user response
        "401":
          $ref: "#/components/responses/UnauthorizedError"
      operationId: GetUser
      tags:
        - Marketplace User
    post:
      security:
        - Marketplace-Authentication: []
      summary: Update user
      description: Updates current marketplace user, inferred from token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Updated marketplace user
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
      operationId: UpdateUser
      tags:
        - Marketplace User
    delete:
      security:
        - Marketplace-Authentication: []
      summary: Request account deletion
      description: |
        Requests deletion of the current marketplace user account.
        This marks the user with a deleteRequestedAt timestamp and sends a notification for manual processing.
      responses:
        "200":
          description: Deletion request submitted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
      operationId: DeleteUser
      tags:
        - Marketplace User
  /v1/marketplace/user/login:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
    post:
      security:
        - Marketplace-Authentication: []
      summary: Login with auth provider
      description: Logs in a marketplace user with Apple or Google Sign-In ID token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialLoginPayload"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Marketplace user
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
      operationId: LoginWithAuthProvider
      tags:
        - Marketplace User
...
  /v1/marketplace/companies/{company_id}/event_types:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
      - name: sort
        in: query
        description: "[Sorting](https://api.noona.is/docs/working-with-the-apis/sorting)"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sort"
      - name: pagination
        in: query
        description: "[Pagination](https://api.noona.is/docs/working-with-the-apis/pagination)"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pagination"
    get:
      summary: List event types
      description: Retrieves all event types for a company.
      parameters:
        - name: company_id
          description: Company ID
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventTypeFilter"
        - name: employee_id
          in: query
          deprecated: true
          description: "Use filter instead"
          schema:
            type: string
        - name: event_type_ids
          in: query
          deprecated: true
          description: "Use filter instead"
          schema:
            type: array
            example: ["te7da9wd8ua28a1ab", "xa7da9wd8ua01a134"]
            items:
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventTypes"
          description: Event types
        "400":
          $ref: "#/components/responses/BadRequestError"
      operationId: ListEventTypes
      tags:
        - Event Types
...
  /v1/marketplace/time_slot_reservations:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
    post:
      security:
        - Marketplace-Authentication: []
      summary: Create time slot reservation
      description: |
        Creates a time slot reservation for a specific event type at a company.

        The reservation can then be used to create an event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeSlotReservation"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeSlotReservation"
          description: Reservation
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTimeSlotReservationError"
      operationId: CreateTimeSlotReservation
      tags:
        - Time Slots
  /v1/marketplace/time_slot_reservations/{time_slot_reservation_id}:
    parameters:
      - $ref: "#/components/parameters/select"
      - $ref: "#/components/parameters/expand"
      - name: time_slot_reservation_id
        in: path
        required: true
        schema:
          type: string
          example: "dwawd8awudawd"
    get:
      security:
        - Marketplace-Authentication: []
      summary: Get time slot reservation
      description: |
        Gets a time slot reservation.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeSlotReservation"
          description: Reservation
      operationId: GetTimeSlotReservation
      tags:
        - Time Slots
    post:
      security:
        - Marketplace-Authentication: []
      summary: Update time slot reservation
      description: |
        Updates a time slot reservation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeSlotReservationUpdate"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeSlotReservation"
          description: Reservation
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTimeSlotReservationError"
      operationId: UpdateTimeSlotReservation
      tags:
        - Time Slots
    delete:
      security:
        - Marketplace-Authentication: []
      summary: Delete time slot reservation
      description: |
        Deletes a time slot reservation.
      responses:
        "200":
          description: Time Slot Reservation deleted
      operationId: DeleteTimeSlotReservation
      tags:
        - Time Slots
...
components:
  securitySchemes:
    Marketplace-Authentication:
      type: apiKey
      in: header
      name: Authorization
      description: |
        The Noona Marketplace API uses an Authorization header with JWT Bearer tokens to authenticate marketplace users.
...
  parameters:
    select:
      name: select
      in: query
      description: "[Field Selector](https://api.noona.is/docs/working-with-the-apis/select)"
      required: false
      schema:
        type: array
        items:
          type: string
    expand:
      name: expand
      in: query
      description: "[Expandable attributes](https://api.noona.is/docs/working-with-the-apis/expandable_attributes)"
      required: false
      schema:
        type: array
        items:
          type: string
    search:
      name: search
      in: query
      description: "[Search](https://api.noona.is/docs/working-with-the-apis/search)"
      required: false
      schema:
        type: string
...
  schemas:
...
    Companies:
      type: array
      items:
        $ref: "#/components/schemas/Company"
...
    EmployeeFilter:
      type: object
      description: "[Filtering](https://api.noona.is/docs/working-with-the-apis/filtering)"
      properties:
        event_type_ids:
          type: array
          description: Filters employees by the event types they are available for.
          items:
            type: string
...
    Employees:
      type: array
      items:
        $ref: "#/components/schemas/Employee"
...
    EventFilter:
      type: object
      description: "[Filtering](https://api.noona.is/docs/working-with-the-apis/filtering)"
      properties:
        from:
          type: string
          format: date-time
          description: Only return events where starts_at is after this timestamp.
        to:
          type: string
          format: date-time
          description: Only return events where starts_at is before this timestamp.
    Events:
      type: array
      items:
        $ref: "#/components/schemas/Event"
...
    EventCreateOverrides:
      type: object
      required:
        - time_slot_reservation
      properties:
        space:
          readOnly: true
        employee:
          readOnly: true
        company:
          readOnly: true
        cancel_reason:
          readOnly: true
        event_types:
          readOnly: true
        status:
          readOnly: true
        starts_at:
          readOnly: true
...
    EventUpdate:
      allOf:
        - $ref: "#/components/schemas/Event"
        - $ref: "#/components/schemas/EventUpdateOverrides"
...
    Sort:
      type: object
      properties:
        field:
          type: string
        order:
          type: string
          enum: [asc, desc]
        ignore_casing:
          type: boolean
          default: false
          description: |
            When true, use case-insensitive sorting for string fields.
            Note: This may prevent the use of indexes for sorting operations impacting performance.
    CompanySort:
      type: object
      properties:
        field:
          type: string
          description: |
            All fields can be used for sorting, along with these special fields:
            - `popular` - Sorts companies by favorite count.
            - `distance` - Sorts companies by distance from the provided location.
            - `onboarded_marketplace` - Sorts companies by when they were onboarded to the marketplace.
            - `enterprise_order` - Sorts companies by their custom order within their enterprise.

            Order is ignored when a special field is used.
        order:
          type: string
          enum: [asc, desc]
...
    Pagination:
      type: object
      properties:
        offset:
          type: integer
          format: int32
          example: 0
          default: 0
        limit:
          type: integer
          format: int32
          example: 10
          default: 500
          maximum: 500
          minimum: 1
          description: The maximum number of results to return.
...
    CreateEventErrorOneOf:
      oneOf:
        - $ref: "#/components/schemas/CreateEventError"
        - $ref: "#/components/schemas/PaymentActionError"
...